requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
imapclient
beautifulsoup4
requests
slack-sdk
openai
langchain
tiktoken
python-dotenv
flask
fastapi
uvicorn
pydantic
httpx
transformers
tensorflow
torch 
torchvision
torchaudio
groq
google-auth-oauthlib 
google-auth-httplib2 
google-auth
